+++ src/Phan/Language/UnionType.php
@@ -17,9 +17,9 @@ use Phan\Language\Type\BoolType;
 use Phan\Language\Type\FalseType;
 use Phan\Language\Type\FloatType;
 use Phan\Language\Type\GenericArrayType;
-use Phan\Language\Type\GenericMultiArrayType;
 use Phan\Language\Type\IntType;
 use Phan\Language\Type\MixedType;
+use Phan\Language\Type\MultiType;
 use Phan\Language\Type\NullType;
 use Phan\Language\Type\StaticType;
 use Phan\Language\Type\TemplateType;
@@ -145,7 +145,7 @@ class UnionType implements \Serializable
                 return Type::fromFullyQualifiedString($type_name);
             }, self::extractTypeParts($fully_qualified_string));
 
-            $unique_types = self::getUniqueTypes(self::normalizeGenericMultiArrayTypes($types));
+            $unique_types = self::getUniqueTypes(self::normalizeMultiTypes($types));
             if (\count($unique_types) === 1) {
                 $union_type = \reset($unique_types)->asUnionType();
             } else {
@@ -213,7 +213,7 @@ class UnionType implements \Serializable
                 $source
             );
         }
-        return UnionType::of(self::normalizeGenericMultiArrayTypes($types));
+        return UnionType::of(self::normalizeMultiTypes($types));
     }
 
     /**
@@ -258,16 +258,16 @@ class UnionType implements \Serializable
     }
 
     /**
-     * Expands any GenericMultiArrayType instances in $types if necessary.
+     * Expands any GenericMultiArrayType and ScalarRawType instances in $types if necessary.
      *
      * @param array<int,Type> $types
      * @return array<int,Type>
      */
-    public static function normalizeGenericMultiArrayTypes(array $types) : array
+    public static function normalizeMultiTypes(array $types) : array
     {
         foreach ($types as $i => $type) {
-            if ($type instanceof GenericMultiArrayType) {
-                foreach ($type->asGenericArrayTypeInstances() as $new_type) {
+            if ($type instanceof MultiType) {
+                foreach ($type->asIndividualTypeInstances() as $new_type) {
                     $types[] = $new_type;
                 }
                 unset($types[$i]);